docker/qwen3_embedder/requirements.txt
flask
llama-cpp-python==0.3.9
